requirements.txt
asgiref==3.7.2
cloudinary==1.40.0
dj-database-url==2.1.0
django==5.0.6
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-compressor==4.3
django-cors-headers==4.3.1
django-csp==3.6
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==2.2.7  
gunicorn==22.0.0
openai==1.35.10
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
setuptools==69.2.0
whitenoise==6.7.0
django-ratelimit==4.1.0
phonenumbers
django-two-factor-auth
django-redis
celery
django-celery-beat
redis